+++ tests/data/full.atom.py
@@ -8,7 +8,7 @@ from reader._types import FeedData
 
 feed = FeedData(
     url=f'{url_base}full.atom',
-    updated=datetime.datetime(2003, 12, 13, 18, 30, 2),
+    updated=datetime.datetime(2003, 12, 13, 18, 30, 2, tzinfo=datetime.timezone.utc),
     title='Example Feed',
     link='http://example.org/',
     author='John Doe',
@@ -20,11 +20,15 @@ entries = [
     EntryData(
         feed_url=feed.url,
         id='urn:uuid:1225c695-cfb8-4ebb-aaaa-80da344efa6a',
-        updated=datetime.datetime(2003, 12, 13, 18, 30, 2),
+        updated=datetime.datetime(
+            2003, 12, 13, 18, 30, 2, tzinfo=datetime.timezone.utc
+        ),
         title='Atom-Powered Robots Run Amok',
         link='http://example.org/2003/12/13/atom03',
         author='John Doe',
-        published=datetime.datetime(2003, 12, 13, 17, 17, 51),
+        published=datetime.datetime(
+            2003, 12, 13, 17, 17, 51, tzinfo=datetime.timezone.utc
+        ),
         summary='Some text.',
         content=(
             # the text/plain type comes from feedparser
@@ -59,7 +63,7 @@ entries = [
     EntryData(
         feed_url=feed.url,
         id='urn:uuid:00000000-cfb8-4ebb-aaaa-00000000000',
-        updated=datetime.datetime(2003, 12, 13, 0, 0, 0),
+        updated=datetime.datetime(2003, 12, 13, 0, 0, 0, tzinfo=datetime.timezone.utc),
         title='Atom-Powered Robots Run Amok Again',
         # link comes from feedparser
         link='urn:uuid:00000000-cfb8-4ebb-aaaa-00000000000',