+++ tests/data/full.json.py
@@ -20,11 +20,11 @@ entries = [
     EntryData(
         feed_url=feed.url,
         id='2',
-        updated=datetime.datetime(2020, 1, 4, 0, 0),
+        updated=datetime.datetime(2020, 1, 4, 0, 0, tzinfo=datetime.timezone.utc),
         title="Title",
         link="https://example.org/second-item",
         author="mailto:joe@example.com",
-        published=datetime.datetime(2020, 1, 2, 21, 0),
+        published=datetime.datetime(2020, 1, 2, 21, 0, tzinfo=datetime.timezone.utc),
         summary="A summary",
         content=(
             Content(
@@ -56,7 +56,7 @@ entries = [
         title=None,
         link='https://example.org/initial-post',
         author='Jane',
-        published=datetime.datetime(2020, 1, 2, 12, 0),
+        published=datetime.datetime(2020, 1, 2, 12, 0, tzinfo=datetime.timezone.utc),
         summary=None,
         content=(
             Content(value='<p>Hello, world!</p>', type='text/html', language='en'),