+++ tests/test_lazy_imports.py
@@ -100,9 +100,9 @@ S_UPDATE_FEEDS_WORKERS = "reader.update_feeds(workers=2)", {
 }
 S_UPDATE_SEARCH = """\
 from reader._types import EntryData, EntryUpdateIntent
-from datetime import datetime
+from datetime import datetime, timezone
 reader.add_feed('one', allow_invalid_url=True)
-dt = datetime(2010, 1, 1)
+dt = datetime(2010, 1, 1, tzinfo=timezone.utc)
 entry = EntryData('one', 'entry', summary='summary')
 reader._storage.add_or_update_entry(EntryUpdateIntent(entry, dt, dt, dt, dt))
 reader.update_search()