+++ tests/test_plugins_mark_as_read.py
@@ -3,7 +3,6 @@ import os
 
 import pytest
 from fakeparser import Parser
-from utils import naive_datetime
 from utils import utc_datetime as datetime
 
 
@@ -25,7 +24,7 @@ def test_regex_mark_as_read(make_reader):
     one = parser.feed(1, datetime(2010, 1, 1))
     parser.entry(1, 1, datetime(2010, 1, 1), title='match old')
 
-    reader._now = lambda: naive_datetime(2010, 1, 1)
+    reader._now = lambda: datetime(2010, 1, 1)
     reader.add_feed(one)
 
     reader.update_feeds()
@@ -39,7 +38,7 @@ def test_regex_mark_as_read(make_reader):
     two = parser.feed(2, datetime(2010, 1, 2))
     parser.entry(2, 3, datetime(2010, 1, 2), title='match other')
 
-    reader._now = lambda: naive_datetime(2010, 2, 1)
+    reader._now = lambda: datetime(2010, 2, 1)
     reader.add_feed(two)
     reader.update_feeds()
 
@@ -50,7 +49,7 @@ def test_regex_mark_as_read(make_reader):
 
     parser.entry(1, 3, datetime(2010, 1, 2), title='no match once again')
 
-    reader._now = lambda: naive_datetime(2010, 3, 1)
+    reader._now = lambda: datetime(2010, 3, 1)
     reader.update_feeds()
 
     assert get_entry_data(read=True) == {