+++ tests/test_reader_private.py
@@ -5,8 +5,8 @@ from datetime import datetime
 import pytest
 from fakeparser import Parser
 from fakeparser import ParserThatRemembers
-from utils import naive_datetime
 from utils import utc_datetime
+from utils import utc_datetime as datetime
 
 from reader import Entry
 from reader import EntryNotFoundError
@@ -26,8 +26,6 @@ def test_update_stale(reader, update_feed, entry_updated):
     regardless of their .updated or caching headers.
 
     """
-    from utils import utc_datetime as datetime
-
     parser = ParserThatRemembers()
     parser.http_etag = 'etag'
     parser.http_last_modified = 'last-modified'
@@ -41,7 +39,7 @@ def test_update_stale(reader, update_feed, entry_updated):
 
     reader.add_feed(feed.url)
 
-    reader._now = lambda: naive_datetime(2010, 1, 1)
+    reader._now = lambda: datetime(2010, 1, 1)
     update_feed(reader, feed.url)
 
     assert {(f.url, f.title, f.last_updated) for f in reader.get_feeds()} == {
@@ -62,7 +60,7 @@ def test_update_stale(reader, update_feed, entry_updated):
 
     if entry_updated:
         # nothing changes after update
-        reader._now = lambda: naive_datetime(2010, 1, 2)
+        reader._now = lambda: datetime(2010, 1, 2)
         update_feed(reader, feed.url)
         assert {(f.url, f.title, f.last_updated) for f in reader.get_feeds()} == {
             (feed.url, feed.title, datetime(2010, 1, 1))
@@ -74,7 +72,7 @@ def test_update_stale(reader, update_feed, entry_updated):
     # but it does if we mark the feed as stale
     parser.calls[:] = []
     reader._storage.set_feed_stale(feed.url, True)
-    reader._now = lambda: naive_datetime(2010, 1, 3)
+    reader._now = lambda: datetime(2010, 1, 3)
     update_feed(reader, feed.url)
     assert parser.calls == [(feed.url, None, None)]
     assert {(f.url, f.title, f.last_updated) for f in reader.get_feeds()} == {
@@ -87,8 +85,6 @@ def test_update_stale(reader, update_feed, entry_updated):
 
 def test_update_parse(reader, update_feed):
     """Updated feeds should pass caching headers back to ._parser()."""
-    from utils import utc_datetime as datetime
-
     parser = ParserThatRemembers()
     parser.http_etag = 'etag'
     parser.http_last_modified = 'last-modified'
@@ -118,8 +114,6 @@ def test_delete_entries(reader):
     and it will be called by plugins.
 
     """
-    from utils import utc_datetime as datetime
-
     reader._parser = parser = Parser()
     feed = parser.feed(1, datetime(2010, 1, 1))
     entry = parser.entry(1, 1, datetime(2010, 1, 1))