+++ CHANGES.rst
@@ -13,6 +13,9 @@ Unreleased
 
 * Stop using deprecated :mod:`sqlite3` datetime converters/adapters.
   (:issue:`321`)
+* Change all :doc:`internal APIs <internal>` to use timezone-aware datetimes,
+  with the timezone set to UTC.
+  (:issue:`321`)
 * In the API documentation,
   fall back to type hints if hand-written parameter types are not available.
   Add relevant :ref:`documentation` guidelines to the dev documentation.